Refactoring

+++ refactoring/src/main/java/com/refactor/helper/IRentCalculationHelper.java
@@ -6,5 +6,5 @@ import com.refactor.model.Rental;
  * Created by abhimanyus on 2/5/18.
  */
 public interface IRentCalculationHelper {
-    int calculate(Rental rental);
+    double calculate(Rental rental);
 }
